+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.1.0] - 2020-02-03
+
+### Added
+
+- Add rectangle installation instructions
+- Add licecap installation instructions
+- Add skitch installation instructions
+- Add daisydisk installation instructions
+- Add pdf-expert installation instructions
+- Add flycut installation instructions
+- Add download github subfolder alias
+
 ## [6.0.0] - 2020-01-20
 
 ### Added
@@ -302,6 +314,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add file for bash aliases
 - Documentation.
 
+[6.1.0]: https://github.com/trystan2k/dotfiles/compare/v6.0.0...v6.1.0
 [6.0.0]: https://github.com/trystan2k/dotfiles/compare/v5.10.0...v6.0.0
 [5.10.0]: https://github.com/trystan2k/dotfiles/compare/v5.9.0...v5.10.0
 [5.9.0]: https://github.com/trystan2k/dotfiles/compare/v5.8.0...v5.9.0